+++ adapters/PubMatic/CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Next version
 - Supports waterfall interstitial ad format.
 - Supports waterfall rewarded ad format.
+- Supports waterfall banner ad format.
 
 #### Version 4.8.1.0 (In progress)